+++ behave_test_automation/pah_behave/features/selectors/base_selectors.py
@@ -1,0 +1,4 @@
+from selenium.webdriver.common.by import By
+
+#main_menu
+hamburger_menu = (By.CSS_SELECTOR, '.bm-burger-button')
