+++ behave_test_automation/pah_behave/features/selectors/login_selectors.py
@@ -6,6 +6,3 @@ password_field = (By.CSS_SELECTOR, 'input[name="password"]')
 
 #login button
 login_button_enabled = (By.CSS_SELECTOR, '.btn-primary')
-
-#main_menu
-hamburger_menu = (By.CSS_SELECTOR, '.bm-burger-button')
