+++ behave_test_automation/pah_behave/features/steps/common_actions.py
@@ -1,0 +1,11 @@
+from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.support.wait import WebDriverWait
+
+
+def short_wait(web_driver, selector) -> None:
+    """
+    Wait for an element to appear in browser. Time value configurable in storage.py
+    """
+    element_short_wait = expected_conditions.visibility_of_element_located(selector)
+    browser_wait = WebDriverWait(web_driver, 10)
+    browser_wait.until(element_short_wait)
